# Fix for mir_performance_tests failure due to not cleaning up our displays