--- a/src/platforms/eglstream-kms/server/platform.h
+++ b/src/platforms/eglstream-kms/server/platform.h
@@ -65,6 +65,8 @@ public:
         EGLDeviceEXT device,
         std::shared_ptr<DisplayReport> display_report);
 
+    ~DisplayPlatform();
+
     auto create_display(
         std::shared_ptr<DisplayConfigurationPolicy> const& initial_conf_policy,
         std::shared_ptr<GLConfig> const& gl_config)