## How to test
- On an nvidia + intel platform, run:
```sh
sudo MIR_EXPERIMENTAL_HYBRID_GRAPHICS=1 MIR_SERVER_PLATFORM_RENDERING_LIBS=mir:eglstream-kms XDG_RUNTIME_DIR=/run/ ./bin/mir_performance_tests --gtest_filter=GLMark2Wayland.*
```

## Expected output
```
LD_LIBRARY_PATH=./bin/../lib/
MIR_SERVER_PLATFORM_PATH=./bin/../lib/server-modules/
exec=./bin/mir_performance_tests.bin
Running main() from main.cpp
Note: Google Test filter = GLMark2Wayland.*
[==========] Running 2 tests from 1 test suite.
[----------] Global test environment set-up.
[----------] 2 tests from GLMark2Wayland
[ RUN      ] GLMark2Wayland.fullscreen
Saving server logs to: /tmp/GLMark2Wayland_fullscreen_server.log
libEGL warning: egl: failed to create dri2 screen
libEGL warning: egl: failed to create dri2 screen
libEGL warning: egl: failed to create dri2 screen
libEGL warning: egl: failed to create dri2 screen
Saving GLMark2 detailed results to: /tmp/GLMark2Wayland_fullscreen.log
[       OK ] GLMark2Wayland.fullscreen (11871 ms)
[ RUN      ] GLMark2Wayland.windowed
Saving server logs to: /tmp/GLMark2Wayland_windowed_server.log
libEGL warning: egl: failed to create dri2 screen
libEGL warning: egl: failed to create dri2 screen
libEGL warning: egl: failed to create dri2 screen
libEGL warning: egl: failed to create dri2 screen
Saving GLMark2 detailed results to: /tmp/GLMark2Wayland_windowed.log
[       OK ] GLMark2Wayland.windowed (11712 ms)
[----------] 2 tests from GLMark2Wayland (23584 ms total)

[----------] Global test environment tear-down
[==========] 2 tests from 1 test suite ran. (23584 ms total)
[  PASSED  ] 2 tests.
```