@RAOF 2023-10-16
> For clarity, should we be testing against `EGL_NO_STREAM_KHR` (for example)?

Yeah. I can't find it offhand, but I'm *pretty* sure that there's at least one `EGL_NO_FOO` that's *not* 0.