@mattkae 2023-10-16
> For clarity, should we be testing against `EGL_NO_STREAM_KHR` (for example)?
> 
> And this looks like the same issue exists on main and maybe better applied there?

We are already checking against `EGL_NO_STREAM_KHR` at construction. Unless I am misunderstanding you?

The original error occurs in `DisplayBuffer::post()` at

```
if (nv_stream(dpy).eglStreamConsumerAcquireAttribNV(dpy, output_stream, acquire_attribs) != EGL_TRUE)
...
```